[#79] New assumptions and ad-hoc solution for run finding
There are some tricky mazes where the current Implementation of the
shortest path does not work, so I created a simple alternative
bfs_shortest_path that will find the shortest path without edge weights.
The crucial insight that made this simplification possible was the fact
that during the speedy run we should no longer make any assumptions that
could shoot us in the foot. I created a method finalize_discovery to
label any wall position we did not gain any information on during the
discovery runs as containing a wall.